refactor: corrige path dos services
